Fix Chrome navigation bug
Some time ago, Blink/WebKit had a bug which caused a popstate event to
be fired when a document was loaded the first time. This was worked
around in appchan-x by ignoring the first popstate event. Meanwhile the
browser bug has been fixed and the workaround has itself turned into a
bug. This commit removes the workaround.

Upstream issue:
https://code.google.com/p/chromium/issues/detail?id=63040

Closes #885
Closes #829
Closes #934